+++ app/assets/stylesheets/partials/_items.scss
@@ -88,109 +88,98 @@
         font-size: 0.8rem;
       }
 
-      .dropzone-container, .dropzone-container2 {
-        // border: solid 2px black;
-        // display: flex;
-        height: 210px;
-        // margin: 1em 0;
-        // padding: 70px 155px;
+      &__dropzone2 {
+        display: none;
+      }
+
+      &__dropzone, &__dropzone2 {
+        height: 162px;
         position: relative;
         cursor: pointer;
-      }
+        margin-top: 16px;
 
-      #preview, #preview2 {
-        // border: solid 2px yellow;
-        // display: inline-block;
-        float: left;
-        // width: 100%;
-      }
-      .dropzone-area, .dropzone-area2 {
-        border: solid 2px red;
-        // display: inline-block;
-        border: dashed 0.2px rgb(207, 206, 206);
-        background-color: #f5f5f5;
+        #preview, #preview2 {
+          float: left;
+          height: 100%;
 
+          // プレビュー画像のスタイル
+          .add_img {
+            float: left;
+            height: 100%;
+            margin: 0 5px;
+            background: rgb(245, 245, 245);
+            position: relative;
+            border: solid 1px $border_gray;
+          }
+    
+          .img_area {
+            width: 111px;
+            height: 116px;
+            display: flex;
+            justify-content: center;
+            align-items: center;
+          }
+    
+          .image {
+            width:auto;
+            height:auto;
+            max-width:100%;
+            max-height:100%;
+          }
 
-        float: right;
-        width: 100%;
-        height: 100%;
-        text-align: center;
-        line-height: 210px;
-      }
+          .btn_wrapper {
+            position: absolute;
+            bottom: 0;
 
-      .upload-image {
-        display: none;
-      }
-
-      .dropzone-container2 {
-        display: none;
+            .btn_edit, .btn_delete {
+              display: inline-block;
+              width: 55px;
+              background: #f5f5f5;
+              line-height: 44px;
+              text-align: center;
+              color: #0099e8;
+              text-decoration: none;
+            }
       
-      }
-
-      a {
-        text-align: center;
-
-        font: {
-          weight: 600;
-          size: 0.8rem;
-          // family: sans-serif;
+            .btn_edit {
+              border-right: solid 1px $border_gray;
+            }
+          }
+          // ここまでプレビュー画像のスタイル
         }
 
-        .top {
-          padding-left: 30px;
+        &__input, &__input2 {
+          border: dashed 0.2px rgb(207, 206, 206);
+          background-color: #f5f5f5;
+          float: right;
+          width: 100%;
+          height: 100%;
+          text-align: center;
+          position: relative;
+
+          .upload-image {
+            display: none;
+          }
+
+          &__description {
+            height: 100%;
+            width: 100%;
+            cursor: pointer;
+            font: {
+              weight: 600;
+              size: 0.8rem;
+            }
+    
+            &__text {
+              position: absolute;
+              top: 50%;
+              left: 50%;
+              -webkit-transform : translate(-50%,-50%);
+              transform : translate(-50%,-50%);
+            }
+          }
         }
       }
-
-      // #preview2 {
-      //   border: solid 2px green;
-      //   // display: inline-block;
-      //   float: left;
-      //   // width: 100%;
-      // }
-
-      // .dropzone-area2 {
-      //   border: solid 2px blue;
-      //   float: right;
-      //   width: 100%;
-      //   height: 100%;
-      //   line-height: 210px;
-
-      //   // display: none;
-      // }
-
-
-      // .drop-box {
-      //   height: 70px;
-      //   margin: 1em 0;
-      //   background-color: #f5f5f5;
-      //   border: dashed 0.2px rgb(207, 206, 206);
-      //   padding: 70px 155px;
-      //   position: relative;
-      //   cursor: pointer;
-
-      //   &__discription {
-      //     text-align: center;
-      //     position: absolute;
-      //     top: 50%;
-      //     left: 50%;
-      //     -webkit-transform : translate(-50%,-50%);
-      //     transform : translate(-50%,-50%);
-
-      //     a {
-      //       text-align: center;
-
-      //       font: {
-      //         weight: 600;
-      //         size: 0.8rem;
-      //         family: sans-serif;
-      //       }
-
-      //       .top {
-      //         padding-left: 30px;
-      //       }
-      //     }
-      //   }
-      // }
     }
 
     &__info {
