+++ app/controllers/items_controller.rb
@@ -14,9 +14,10 @@ class ItemsController < ApplicationController
 
   def create
     @item = Item.create(item_params)
+    binding.pry
     respond_to do |format|
       if @item.save
-        binding.pry
+        # binding.pry
         params[:item_images][:image].each do |image|
           @item.item_images.create(image_url: image, item_id: @item.id)
         end
@@ -51,7 +52,7 @@ class ItemsController < ApplicationController
 
   private
   def item_params
-    params.require(:item).permit(:name, :text, :category_id, :size_id, :brand_id, :condition, :delivery_fee_payer, :delivery_type, :delibery_from_area, :delivery_days, :price, item_images_attributes: [:id, :image_url, :item_id]).merge(seller_id: 1)
+    params.require(:item).permit(:name, :text, :category_id, :size_id, :brand_id, :condition, :delivery_fee_payer, :delivery_type, :delibery_from_area, :delivery_days, :price, item_images_attributes: [:id, :image_url, :item_id])
     #### ログイン機能ができたら.merge(seller_id: current_user.id)
   end
