+++ app/models/item.rb
@@ -1,8 +1,5 @@
 class Item < ApplicationRecord
 
-  belongs_to :buyer, class_name: 'User', foreign_key: 'buyer_id', optional: true
-  belongs_to :seller, class_name: 'User', foreign_key: 'seller_id'
-
   belongs_to :category
 
   has_many :item_images, dependent: :destroy
@@ -10,5 +7,5 @@ class Item < ApplicationRecord
   has_one :deal
   belongs_to :category
   belongs_to :size
-  belongs_to :brand
+  belongs_to :brand, optional: true
 end
