+++ db/seeds.rb
@@ -2,4 +2,4 @@ require './db/seeds/brands.rb'
 require './db/seeds/sizes.rb'
 require './db/seeds/categories.rb'
 require './db/seeds/size_categories.rb'
-require './db/seeds/size_users.rb'
+require './db/seeds/users.rb'
